project11/requirements.txt
paramiko
python-dotenv
sshtunnel
fastapi
hypercorn
pydantic
python-multipart
requests
numpy
openai
opencv-python
tqdm
